requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
Django==3.2.13
django-appconf==1.0.5
django-autocomplete-light==3.9.1
django-cors-headers==3.13.0
django-cryptography==1.0
django-el-pagination==3.3.0
django-filter==22.1
django-mail-queue==3.2.5
django-registration-redux==2.10
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==4.6.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
numpy==1.21.6
openpyxl==3.0.10
openpyxl-image-loader==1.0.5
packaging==22.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.5
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2022.1
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.10
xlrd==1.2.0
xlwt==1.3.0
zipp==3.8.1